dir errors

+++ mkc_dirop.c
@@ -18,6 +18,7 @@
 #include <fcntl.h>
 #include <unistd.h>
 #include <wchar.h>
+#include <errno.h>
 
 #if __has_include (<windows.h>)
 # define WIN32_LEAN_AND_MEAN 1
@@ -26,36 +27,48 @@
 
 #include "mkc_def.h"
 #include "mkc_dirop.h"
+#include "mkc_error.h"
 #include "mkc_fileop.h"
 #include "mkc_osdir.h"
 #include "mkc_string.h"
 
-static int mkc_dirop_make_recursive (const char *dirname);
-static int mkc_dirop_makedir (const char *dirname);
+static int mkc_dirop_make_recursive (const char *dirname, mkc_error_t *mkcerr);
+static int mkc_dirop_makedir (const char *dirname, mkc_error_t *mkcerr);
 
 int
-mkc_dirop_make (const char *dirname)
+mkc_dirop_make (const char *dirname, mkc_error_t *mkcerr)
 {
   int rc;
-  rc = mkc_dirop_make_recursive (dirname);
+  rc = mkc_dirop_make_recursive (dirname, mkcerr);
   return rc;
 }
 
 /* deletes the directory */
 /* if the directory only contains other dirs, the empty dirs will */
 /* be recursively deleted */
-bool
-mkc_dirop_delete (const char *dirname, int flags)
+int
+mkc_dirop_delete (const char *dirname, int flags, mkc_error_t *mkcerr)
 {
   dirhandle_t   *dh;
   char          *fname;
   char          temp [MKC_PATH_MAX];
 
+  if (mkc_is_directory (dirname)) {
+    /* not an error */
+    return MKC_OK;
+  }
+
   if (! mkc_is_directory (dirname)) {
-    return false;
+    mkc_error_set (mkcerr, MKC_ERR_DIR_NOT_A_DIR, 0, dirname);
+    return MKC_ERR_FAILURE;
   }
 
   dh = mkc_osdir_open (dirname);
+  if (dh == NULL) {
+    mkc_error_set (mkcerr, MKC_ERR_DIR_UNABLE_TO_OPEN, errno, dirname);
+    return MKC_ERR_FAILURE;
+  }
+
   while ((fname = mkc_osdir_iterate (dh)) != NULL) {
     if (strcmp (fname, ".") == 0 ||
         strcmp (fname, "..") == 0) {
@@ -66,16 +79,16 @@ mkc_dirop_delete (const char *dirname, int flags)
     snprintf (temp, sizeof (temp), "%s/%s", dirname, fname);
     if (fname != NULL) {
       if (mkc_is_directory (temp)) {
-        if (! mkc_dirop_delete (temp, flags)) {
+        if (! mkc_dirop_delete (temp, flags, mkcerr)) {
           free (fname);
           mkc_osdir_close (dh);
-          return false;
+          return MKC_ERR_FAILURE;
         }
       } else {
         if ((flags & DIROP_ONLY_IF_EMPTY) == DIROP_ONLY_IF_EMPTY) {
           free (fname);
           mkc_osdir_close (dh);
-          return false;
+          return MKC_ERR_FAILURE;
         }
         mkc_file_delete (temp);
       }
@@ -93,6 +106,10 @@ mkc_dirop_delete (const char *dirname, int flags)
     wchar_t       * tdirname;
 
     tdirname = mkc_towide (dirname);
+    if (tdirname == NULL) {
+      mkc_error_set (mkcerr, MKC_ERR_OUT_OF_MEMORY, 0, NULL);
+      return MKC_ERR_FAILURE;
+    }
     RemoveDirectoryW (tdirname);
     free (tdirname);
   }
@@ -100,16 +117,22 @@ mkc_dirop_delete (const char *dirname, int flags)
   rmdir (dirname);
 #endif
 
-  return true;
+  return MKC_OK;
 }
 
 /* internal routines */
 
 static int
-mkc_dirop_make_recursive (const char *dirname)
+mkc_dirop_make_recursive (const char *dirname, mkc_error_t *mkcerr)
 {
   char    tbuff [MKC_PATH_MAX];
   char    *p = NULL;
+  int     rc = MKC_OK;
+
+  if (mkc_is_directory (dirname)) {
+    /* not an error */
+    return MKC_OK;
+  }
 
   stpecpy (tbuff, tbuff + MKC_PATH_MAX, dirname);
   mkc_trim_char (tbuff, '/');
@@ -117,29 +140,41 @@ mkc_dirop_make_recursive (const char *dirname)
   for (p = tbuff + 1; *p; p++) {
     if (*p == '/') {
       *p = '\0';
-      mkc_dirop_makedir (tbuff);
+      rc = mkc_dirop_makedir (tbuff, mkcerr);
+      if (rc != 0) {
+        break;
+      }
       *p = '/';
     }
   }
-  mkc_dirop_makedir (tbuff);
-  return 0;
+  if (rc == 0) {
+    rc = mkc_dirop_makedir (tbuff, mkcerr);
+  }
+  return rc;
 }
 
 static int
-mkc_dirop_makedir (const char *dirname)
+mkc_dirop_makedir (const char *dirname, mkc_error_t *mkcerr)
 {
-  int   rc = 1;
+  int   rc = MKC_ERR_FAILURE;
 
 /* windows */
 #if _arg_count_mkdir == 1
   wchar_t   *tdirname = NULL;
 
   tdirname = mkc_towide (dirname);
+  if (tdirname == NULL) {
+    mkc_error_set (mkcerr, MKC_ERR_OUT_OF_MEMORY, 0, NULL);
+    return MKC_ERR_FAILURE;
+  }
   rc = _wmkdir (tdirname);
   free (tdirname);
 #endif
 #if _arg_count_mkdir == 2 && _define_S_IRWXU
   rc = mkdir (dirname, S_IRWXU);
 #endif
+  if (rc != 0) {
+    mkc_error_set (mkcerr, MKC_ERR_DIR_NOT_CREATED, errno, dirname);
+  }
   return rc;
 }
